Cleaner shimming of log
